requirements-suno.txt
# Suno-only pipeline (minimal dependencies)
replicate>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.31.0
google-cloud-storage>=2.16.0
